Clean optuna visualize code
